refactor: re-structure repo and tests
