+++ src/football_pipeline/pipe_001_raw.py
@@ -6,10 +6,7 @@ from football_pipeline.adapters.io_wrapper import FileType, IOWrapper
 from football_pipeline.adapters.logger import RealLogger
 from football_pipeline.adapters.repo import Repo
 from football_pipeline.adapters.time import new_guid, time_now
-from football_pipeline.io import (
-    convert_data_to_df,
-)
-from football_pipeline.transform import add_ingestion_columns
+from football_pipeline.domain.transform import add_ingestion_columns, convert_data_to_df
 
 
 def run_raw_layer(config_path: str, repo: Repo) -> dict:
